refactor(mcp): extract Figma Plugin API scripts to scripts/index.ts
Phase 2 of McpServer.ts decomposition. Move all 11 executeJavaScript
template literals (~900 lines) verbatim into src/main/MCP/scripts/,
imported back into McpServer. No behavior change. McpServer.ts drops
from ~2411 to ~1451 lines.
